moved test scraper container to container folder
