+++ frontend/src/components/App/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import TestScraperFormContainer from '../TestScraperForm/TestScraperFormContainer';
+import TestScraperFormContainer from '../../containers/TestScraperFormContainer';
 import {
 	SignInPage,
 	SignUpPage,
